addressing Matti's comments
